+++ NEWS.md
@@ -1,7 +1,13 @@
+# Version 0.2.4
+* Fixed the code to accept raster integer data type.
+* Added a scaling factor (`scale` arg) to convert condition data to the 0–1 range.
+* The output array/map is now cropped to the supplied polygon mask.
+* Made several improvements on the Rust side, such as using `Arc<T>` for cheap cloning and introducing the `Graph` struct for more organised code.
+
 # Version 0.2.3
 * Fixed isolated pixel calculation using their own values for a synthetic neighbour.
 * Increased the capacity of the graph to take more neighbours for high-resolution rasters.
-* Changed the default value ofthe  `sigma` parameter to 1.0.
+* Changed the default value of the `sigma` parameter to 1.0.
 
 # Version 0.2.2
 * Removed the dependency on GDAL, replacing it with `rasterio`.